Brought back the sampling options
Because sometimes the network is being naughty, so we need to send less
points :cry:
Choose it in the sensor client. Splats in the application that requested
the data will be adapted to fill the whole thing. If one wants to use
different sampling levels for each camera, it looks better if you
disable this (simply remove the _texScale multiplying _Size in the
shader)